infra/requirements.txt
aws-cdk-lib==2.83.0
# constructs>=10.0.0,<11.0.0
python-dotenv==1.0.0
aws-cdk.aws-apigatewayv2-alpha==2.83.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.83.0a0
aws-cdk.aws-apigatewayv2-authorizers-alpha==2.83.0a0
aws-cdk.aws-lambda-python-alpha==2.83.0a0
cdk8s
cdk8s-plus-26

# trio 0.22.0 requires idna, which is not installed.
# trio 0.22.0 requires outcome, which is not installed.
# trio-websocket 0.9.2 requires async-generator>=1.10, which is not installed.
# webdriver-manager 3.8.5 requires packaging, which is not installed.
